add current limits

+++ src/main/java/frc/robot/subsystems/Elevator/ElevatorConstants.java
@@ -37,7 +37,7 @@ public class ElevatorConstants {
       int[] canIds,
       boolean[] reversed,
       double gearRatio,
-      double currentLimit,
+      int currentLimit,
       GravityType gravity,
       EncoderType encoderType,
       int encoderID,
